requirements.txt
requests>=2.31.0
pandas>=2.0.0
schedule>=1.2.0
yfinance>=0.2.36
python-dotenv>=1.0.1
